chore: move projets to packages
