+++ packages/example-inventory/001-hello-world/README.md
