+++ packages/example-inventory/001-hello-world/index.js
