+++ packages/example-inventory/001-hello-world/package.json
