+++ packages/example-inventory/002-connecting-ct/README.md
