+++ packages/example-inventory/002-connecting-ct/config.json
