+++ packages/example-inventory/002-connecting-ct/index.js
