+++ packages/example-inventory/002-connecting-ct/package.json
