+++ packages/example-inventory/README.md
