+++ packages/extension-base/ctp-extension-gcf-adapter.js
@@ -1,14 +1,14 @@
-// Azure
+// Google Cloud Functions
 
 module.exports = {
   ctpExtensionAdapter: function(fn) {
-    return function(context, req) {
-      fn(req.body, ctpResponse(context), context.log);
+    return function(req, res) {
+      fn(req.body, ctpResponse(res), console.log);
     }
   }
 }
 
-function ctpResponse(context) {
+function ctpResponse(res) {
   return {
     reject: createReject(context),
     accept: createAccept(context),
@@ -17,23 +17,14 @@ function ctpResponse(context) {
 
 const createReject = context => (error) => {
   var errorArray = Array.isArray(error) ? error : [error]
-  context.res = {
-    status: 400,
-    body : {
-      errors : [errorArray]
-    }
-  };
-  context.done();
+  res.status(400).json({
+    errors : errorArray
+  });
 }
 
 const createAccept = context => (update = []) => {
   var actionsArray = Array.isArray(update) ? update : [update]
-  context.res = {
-    status: 200,
-    body : {
-      actions : actionsArray
-    }
-  };
-  context.done();
+  res.status(200).json({
+    actions : actionsArray
+  });
 }
-
