+++ packages/extension-base/package.json
@@ -2,4 +2,3 @@
   "name": "sample-extension",
   "version": "1.0.0"
 }
-
